requirements.txt
langchain-astradb==0.3.5
langchain==0.2.14
langchain-core==0.2.38
langchain-community==0.2.10
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
datasets
pypdf==4.3.1
beautifulsoup4
python-dotenv
pandas
flask

-e .
